refactor: split main.ts into focused modules
Moves uploadToS3, getOrCreateFunctionId, createFunctionVersion, and
resolveLatestLockboxVersions into src/upload.ts, src/function/create.ts,
src/function/version.ts, and src/lockbox.ts, and input reading into
readInputs() beside the ActionInputs type. main.ts keeps credential
resolution and orchestration.

Pure movement - the characterization snapshot is unchanged.
